+++ packages/reconciliation-core/src/index.ts
@@ -6,3 +6,4 @@ export * from "./run-resolution.js";
 export * from "./uuid-collision-log.js";
 export * from "./v1-serializer.js";
 export * from "./run-capabilities.js";
+export * from "./console-reconciliation-list.js";